Rename whitelist to allowlist.
