+++ lib/aws-xray-sdk/facets/resources/aws_params_allowlist.rb
