+++ lib/aws-xray-sdk/facets/resources/aws_services_allowlist.rb
